Reorganise application structure
