+++ server/src/chat/chat.gateway.ts
@@ -1,19 +1,13 @@
 import { Logger } from '@nestjs/common';
-import {
-    OnGatewayConnection,
-    OnGatewayDisconnect,
-    SubscribeMessage,
-    WebSocketGateway,
-    WebSocketServer,
-} from '@nestjs/websockets';
+import { SubscribeMessage, WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 import { Server, Socket } from 'socket.io';
 import {
-    Client_AuthenticateEventPayload,
     Client_ChatMessagePayload,
     Client_TypingEventPayload,
     Server_ChatMessagePayload,
 } from 'src/shared/chat-event-payloads.model';
 import { SocketEventPayloadAsFnMap } from 'src/shared/event-payload-map.model';
+import { SocketManagerService } from 'src/socket/socket-manager.service';
 import { SocketEvents } from '../shared/socket-events.model';
 import { ChatService } from './chat.service';
 
@@ -22,94 +16,15 @@ export const GROUP_CHAT = 'global group chat';
 export type TypedSocket = Socket<SocketEventPayloadAsFnMap, SocketEventPayloadAsFnMap, SocketEventPayloadAsFnMap>;
 
 @WebSocketGateway()
-export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
-    constructor(private chatService: ChatService) {}
+export class ChatGateway {
+    constructor(private socketManager: SocketManagerService, private chatService: ChatService) {}
 
     private logger: Logger = new Logger('ChatGateway');
     @WebSocketServer() server: Server<SocketEventPayloadAsFnMap, SocketEventPayloadAsFnMap, SocketEventPayloadAsFnMap>;
 
-    async handleConnection(client: TypedSocket) {
-        this.logger.verbose(`Client connected: ${client.id}`);
-
-        const allSockets = await this.server.allSockets();
-        this.chatService.onClientConnected(allSockets);
-
-        client.emit(SocketEvents.SERVER__AUTHENTICATE_PROMPT, null);
-    }
-    async handleDisconnect(client: TypedSocket) {
-        const user = this.onCLientLogout(client);
-        this.logger.verbose(`${user} disconnected`);
-    }
-
-    @SubscribeMessage(SocketEvents.CLIENT__AUTHENTICATE)
-    async handleAuth(client: TypedSocket, { accessToken }: Client_AuthenticateEventPayload) {
-        if (!accessToken) {
-            client.emit(SocketEvents.SERVER__AUTHENTICATE, { authenticated: false });
-            return;
-        }
-
-        const { authenticated, user, chatIds } = await this.chatService.authenticateSocket(client.id, accessToken);
-
-        client.emit(SocketEvents.SERVER__AUTHENTICATE, { authenticated });
-
-        if (!authenticated) {
-            this.logger.verbose(`'${user.username || '[some user]'}' could not authenticate`);
-            return;
-        }
-
-        this.logger.verbose(`'${user.username}' authenticated successfully`);
-        this.chatService.setUserOnline({ userId: user.id, username: user.username, client }, chatIds);
-        this.chatService.logUsersOnline();
-
-        chatIds.forEach(chatId => {
-            this.emitAllOnlineUsers(chatId);
-            client.broadcast.to(chatId).emit(SocketEvents.SERVER__USER_ONLINE_STATUS_EVENT, {
-                user,
-                online: true,
-                chatIds,
-            });
-        });
-    }
-    @SubscribeMessage(SocketEvents.CLIENT__LOGOUT)
-    handleLogout(client: TypedSocket) {
-        const user = this.onCLientLogout(client);
-        this.logger.verbose(`${user} logged out`);
-    }
-
-    onCLientLogout(client: TypedSocket) {
-        const data = this.chatService.onClientLogout(client.id);
-        if (!data) return '[a not authenticated user]';
-        const { loggedOutUser, chatIds } = data;
-
-        chatIds.forEach(chatId => {
-            client.broadcast.to(chatId).emit(SocketEvents.SERVER__TYPING_EVENT, {
-                username: loggedOutUser.username,
-                isTyping: false,
-                chatId,
-            });
-            client.broadcast.to(chatId).emit(SocketEvents.SERVER__USER_ONLINE_STATUS_EVENT, {
-                user: loggedOutUser,
-                online: false,
-                chatIds,
-            });
-            client.leave(chatId);
-
-            this.emitAllOnlineUsers(chatId);
-        });
-
-        return loggedOutUser.username;
-    }
-
-    private emitAllOnlineUsers(chatId: string) {
-        this.server.to(chatId).emit(SocketEvents.SERVER__USERS_ONLINE, {
-            chatId,
-            usersOnline: this.chatService.getUsersOnlineForChat(chatId).map(u => u.username),
-        });
-    }
-
     @SubscribeMessage(SocketEvents.CLIENT__CHAT_MESSAGE)
     async handleChatMessage(client: TypedSocket, { chatId, messageText }: Client_ChatMessagePayload) {
-        const user = this.chatService.getUserOnline(client.id);
+        const user = this.socketManager.getUserOnline(client.id);
         if (!user) return;
         const isUserChatMember = await this.chatService.isUserChatMember(chatId, user.userId);
         if (!isUserChatMember) return;
@@ -125,7 +40,7 @@ export class ChatGateway implements OnGatewayConnection, OnGatewayDisconnect {
 
     @SubscribeMessage(SocketEvents.CLIENT__TYPING_EVENT)
     async handleTypingEvents(client: TypedSocket, { chatId, isTyping }: Client_TypingEventPayload) {
-        const user = this.chatService.getUserOnline(client.id);
+        const user = this.socketManager.getUserOnline(client.id);
         if (!user) return;
         const isUserChatMember = await this.chatService.isUserChatMember(chatId, user.userId);
         if (!isUserChatMember) return;
