+++ server/src/chat/chat.module.ts
@@ -1,14 +1,14 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from 'src/services/prisma.service';
 import { UsersModule } from 'src/users/users.module';
-import { UsersService } from 'src/users/users.service';
 import { ChatGateway } from './chat.gateway';
 import { ChatService } from './chat.service';
 import { ChatController } from './chat.controller';
+import { SocketModule } from 'src/socket/socket.module';
+import { PrismaModule } from 'src/prisma-abstractions/prisma.module';
 
 @Module({
-    imports: [UsersModule],
-    providers: [ChatGateway, ChatService, PrismaService, UsersService],
+    imports: [UsersModule, SocketModule, PrismaModule],
+    providers: [ChatGateway, ChatService],
     controllers: [ChatController],
 })
 export class ChatModule {}
