+++ server/src/socket/socket.gateway.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { SocketGateway } from './socket.gateway';
+
+describe('SocketGateway', () => {
+    let gateway: SocketGateway;
+
+    beforeEach(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            providers: [SocketGateway],
+        }).compile();
+
+        gateway = module.get<SocketGateway>(SocketGateway);
+    });
+
+    it('should be defined', () => {
+        expect(gateway).toBeDefined();
+    });
+});
