+++ server/src/socket/socket.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { UsersModule } from 'src/users/users.module';
+import { SocketManagerService } from './socket-manager.service';
+import { SocketGateway } from './socket.gateway';
+
+@Module({
+    imports: [UsersModule],
+    providers: [SocketGateway, SocketManagerService],
+    exports: [SocketManagerService],
+})
+export class SocketModule {}
