+++ server/src/users/users.service.ts
@@ -10,8 +10,8 @@ import { JwtService } from '@nestjs/jwt';
 import { User } from '@prisma/client';
 import * as bcrypt from 'bcrypt';
 import { ADMIN_PWD } from 'src/constants';
-import { SELECT_user_preview } from 'src/query-helpers';
-import { PrismaService } from 'src/services/prisma.service';
+import { SELECT_user_preview } from 'src/prisma-abstractions/query-helpers';
+import { PrismaService } from 'src/prisma-abstractions/prisma.service';
 import { LoginCredentialsDTO, SignupCredentialsDTO } from './dto/auth-credetials.dto';
 import { UpdatePasswordDTO, UpdateUserDTO } from './dto/update-user.dto';
 import { JwtPayload } from './jwt-payload.interface';
